+++ src/kaskas/io/providers/analogue.hpp
@@ -11,9 +11,9 @@ namespace kaskas::io {
 /// a provider for any sensor that provides a single fp voltage
 class AnalogueSensor : public Provider {
 public:
-    AnalogueSensor(const std::function<double()>& value_f) : _value_f(value_f){};
+    AnalogueSensor(const std::function<float()>& value_f) : _value_f(value_f){};
 
-    double value() const { return _value_f(); }
+    float value() const { return _value_f(); }
 
     std::unique_ptr<prompt::RPCRecipe> rpc_recipe(const std::string_view& recipe_name,
                                                   const std::string_view& root) override {
@@ -27,30 +27,30 @@ public:
     }
 
 private:
-    const std::function<double()> _value_f;
+    const std::function<float()> _value_f;
 };
 
 /// a provider for any actuator that needs a single fp voltage
 class AnalogueActuator : public Provider {
 public:
     struct FunctionMap {
-        const std::function<double()> value_f;
-        const std::function<void(double)> set_value_f;
-        const std::function<void(double, double, k_time_ms)> fade_to_f;
-        const std::function<void(double, k_time_ms)> creep_to_f;
+        const std::function<float()> value_f;
+        const std::function<void(float)> set_value_f;
+        const std::function<void(float, float, k_time_ms)> fade_to_f;
+        const std::function<void(float, k_time_ms)> creep_to_f;
         const std::function<void()> creep_stop_f;
     };
 
     AnalogueActuator(const FunctionMap& map) : _map(map){};
 
-    double value() const { return _map.value_f(); }
-    void set_value(double value) { _map.set_value_f(value); }
+    float value() const { return _map.value_f(); }
+    void set_value(float value) { _map.set_value_f(value); }
 
-    void fade_to(double setpoint, double increment = 0.1, k_time_ms increment_interval = k_time_ms(150)) {
+    void fade_to(float setpoint, float increment = 0.1, k_time_ms increment_interval = k_time_ms(150)) {
         _map.fade_to_f(setpoint, increment, increment_interval);
     }
 
-    void creep_to(double setpoint, k_time_ms travel_time) { _map.creep_to_f(setpoint, travel_time); }
+    void creep_to(float setpoint, k_time_ms travel_time) { _map.creep_to_f(setpoint, travel_time); }
     void creep_stop() { _map.creep_stop_f(); }
     std::unique_ptr<prompt::RPCRecipe> rpc_recipe(const std::string_view& recipe_name, const std::string_view& root) {
         return {};
